requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
 
# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
 
# Validation
pydantic==2.7.1
 
# Config
python-dotenv==1.0.1
 
# Compression middleware (built-in fastapi, but needs starlette explicitly)
starlette==0.37.2
 